Move buttons to NavBar
