--- client/src/components/LogoutButton.tsx
@@ -1,18 +1,0 @@
-import Button from '@mui/material/Button';
-
-import { logoutUser } from '../actions/authActions';
-import { useAppDispatch } from '../app/hooks';
-
-export default function LogoutButton() {
-  const dispatch = useAppDispatch();
-
-  const handleLogout = async () => {
-    dispatch(logoutUser());
-  };
-
-  return (
-    <Button color="inherit" onClick={handleLogout}>
-      Logout
-    </Button>
-  );
-}
