+++ client/src/components/NavBar.tsx
@@ -1,7 +1,8 @@
-import { AppBar, Box, styled, Toolbar, Typography } from '@mui/material';
+import { AppBar, Box, Button, styled, Toolbar, Typography } from '@mui/material';
+import { useNavigate } from 'react-router-dom';
 
-import CreatePostButton from './CreatePostButton';
-import LogoutButton from './LogoutButton';
+import { logoutUser } from '../actions/authActions';
+import { useAppDispatch } from '../app/hooks';
 
 const Offset = styled('div')(({ theme }) => theme.mixins.toolbar);
 
@@ -19,5 +20,28 @@ export default function NavBar() {
       </AppBar>
       <Offset />
     </Box>
+  );
+}
+
+function CreatePostButton() {
+  const navigate = useNavigate();
+  return (
+    <Button color="inherit" onClick={() => navigate('/feed/new')}>
+      New Post
+    </Button>
+  );
+}
+
+function LogoutButton() {
+  const dispatch = useAppDispatch();
+
+  const handleLogout = async () => {
+    dispatch(logoutUser());
+  };
+
+  return (
+    <Button color="inherit" onClick={handleLogout}>
+      Logout
+    </Button>
   );
 }
